feat: refactor task status update logic and extract helper functions for better maintainability
